Split `abstract class NavigationMode` into `interface NavigatinMode` and `abstract class BaseNavigationMode`.
